requirements.txt
Flask==2.3.2
adafruit-blinka
adafruit-circuitpython-neopixel